Fix wachtwoord vergeten

+++ wwwroot/assets/js/Controllers/Nieuwwachtwoord_Controller.js
@@ -6,6 +6,7 @@ import '../config.js';
 import Controller from './Controller.js';
 
 import view from "../Classes/View.js";
+import App from "../Classes/app";
 
 
 
@@ -18,31 +19,37 @@ export default class Nieuwwachtwoord_Controller extends Controller
         document.head.appendChild(script);
 
         document.querySelector(".zoeken").addEventListener("click", event => {
-            let email = localStorage.getItem("email")
-            const wachtwoord = document.getElementById("wachtwoord").value;
+            const query = App.getFromQueryObject();
+            if(query.email) {
+                //let email = localStorage.getItem("email")
+                const wachtwoord = document.getElementById("wachtwoord").value;
+                if(!App.validatePassword(wachtwoord)){
+                    App.addError(wachtwoord,"Wachtwoord voldoet niet aan de wachtwoord eisen,");
+                    App.ShowNotifyError("Wachtwoord veranderen is fout gegaan");
+                    return
+                }
+                const shaObj = new jsSHA("SHA-512", "TEXT", {encoding: "UTF8"});
+                shaObj.update(wachtwoord);
+                const hash = shaObj.getHash("HEX");
 
-            const shaObj = new jsSHA("SHA-512", "TEXT", {encoding: "UTF8"});
-            shaObj.update(wachtwoord);
-            const hash = shaObj.getHash("HEX");
 
-
-
-            FYSCloud.API.queryDatabase('UPDATE users SET password = ? WHERE email = ?', [hash,email])
-                .then(function(data) {
+                FYSCloud.API.queryDatabase('UPDATE users SET password = ? WHERE email = ?', [hash, query.email])
+                    .then(function (data) {
                         document.getElementById("titel").innerHTML = "Je wachtwoord is veranderd"
                         document.getElementById("titel").style.backgroundColor = "green"
                         document.getElementById("error").innerHTML = "Jouw wachtwoord is veranderd, ga terug naar het login scherm en probeer het uit."
                         document.getElementById("error").style.color = " green"
                         window.localStorage.clear();
-                    setTimeout(
-                        function( )
-                        {
-                            window.location.replace("https://dev-is108-3.fys.cloud/#/home")
-                        }, 2700);
-                }).catch(function (reason) {
-                console.log(reason);
+                        setTimeout(
+                            function () {
+                                window.location.replace("https://dev-is108-3.fys.cloud/#/home")
+                            }, 2700);
+                    }).catch(function (reason) {
+                    console.log(reason);
 
-            });
+                });
+            }
+            App.ShowNotifyError('De url is niet juist voor het veranderen van je wachtwoord');
         })
     }
     render() {
